Add Traversable Flag to Locations
This way I can mark any location as whether or not a player can even travel to it.  Useful for filtering out Regions, Continents, etc. that are too abstract to visit, or a location that's useful to describe to a player but not necessarily accessible (i.e. dangerous terrain).

By default, only LocalAreas and Points of Interest will be traversable unless otherwise marked during generation.